final touches on the remote data source

+++ on-boarding/product_8/lib/features/Product/data/models/product_model.dart
@@ -13,7 +13,7 @@ class ProductModel extends ProductEntity {
         id: json['id'],
         name: json['name'],
         description: json['description'],
-        price: json['price'],
+        price: json['price'].toDouble(),
         imageUrl: json['imageUrl'],
       );
 
@@ -33,3 +33,4 @@ class ProductModel extends ProductEntity {
         imageUrl: imageUrl,
       );
 }
+
